IGNITE-24847 Move some U methods, IgniteException, IgniteCheckedException to ignite-commons
