+++ modules/commons/src/main/java/org/apache/ignite/IgniteException.java
